feat: implement jotai for product state management and refactor App component
- Integrated jotai for managing product state through atoms, enhancing modularity and maintainability.
- Refactored useProducts hook to utilize jotai atoms for product operations, including addition, update, and deletion.
- Renamed App component to AppContent for clarity and wrapped it with a Provider to enable jotai's context.
- Introduced a new productStore file to define product-related atoms, streamlining state management.
